% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/act24_api.R
\name{Act24_API_Wrapper}
\alias{Act24_API_Wrapper}
\alias{getStudies}
\alias{getParticantList}
\alias{getDetailReportQC}
\title{Call Act24 API}
\usage{
getStudies()

getParticantList(studyId)

getDetailReportQC(studyId)
}
\arguments{
\item{studyId}{The studyId assigned by act24. Found in the getStudies results}
}
\value{
a tibble that depends on the api called
}
\description{
These function wrap the act24 api and return a tibble containing the results.


 \strong{getStudies()} returns a tibble containing all the studies that you can
  access. It contains the studyId, which is required to lookup other
  information about an individual staudy.

 \strong{getParticantList(studyId)} returns a tibble of participants for a study

 \strong{getDetailReportQC(studyId)} returns the detailed tibble of activities for a study
}
