% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{act24_auth}
\alias{act24_auth}
\alias{act24_deauth}
\title{Provide API Linkage Id/password}
\usage{
act24_auth(api_id)

act24_deauth()
}
\arguments{
\item{api_id}{The ACT24 API Linkage id}
}
\description{
act24_auth sets the api linkage id (\strong{not the ACT24 research id}) and password.
The password is cached until you close your R session
}
\details{
act24_deauth removes the api linkage information.
}
